syntaxes: update functions regex
Update the functions regex in syntaxes/yagpdbcc.tmLanguage.json.

I ran lytfs[0] and replaced the newlines with the `|` separator,
therefore this regex is in essence generated. In doing so, the regex is
also easier to (manually) maintain—we no longer have weird grouping to
deal with.

It may be worthwhile to consider automating this process, that is,
providing a build script and running that on a regular interval—GitHub
Actions supports a `schedule` trigger that is checked against a crontab.

However, as I am not familiar with the build processes pertaining to
creating VSCode extensions, nor the (admittedly, rather small) codebase,
I'm afraid I'm unable to provide help in more than "check this out" kind
of capacity.

[0]: https://github.com/jo3-l/yagfuncdata

Signed-off-by: Luca Zeuch <l-zeuch@email.de>